+++ pyodide_pack/js/discovery.js
@@ -4,6 +4,9 @@ async function main() {
   let fs = await import("fs");
 
   let pyodide = await loadPyodide()
+  let initSysModules = pyodide.runPython(
+	"import sys; {name: getattr(mod, '__file__', None) for name, mod in sys.modules.items()}"
+  ).toJs({dict_converter : Object.fromEntries});
   let file_list = [];
   const open_orig = pyodide._module.FS.open;
   // Monkeypatch FS.open
@@ -37,7 +40,9 @@ async function main() {
   obj.opened_file_names = file_list;
   obj.loaded_packages = pyodide.loadedPackages;
   obj.find_object_calls = findObjectCalls;
-
+  obj.init_sys_modules = initSysModules;
+  // For some reason there is a double / in the path
+  obj.stdlib_prefix = pyodide.pyimport('sysconfig').get_path('stdlib').replace('//', '/');
   let jsonString = JSON.stringify(obj);
   let file = fs.createWriteStream("{{ output_path }}");
   file.write(jsonString);